pc-controller/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autotest"
version = "1.0.0"
description = "Industrial-grade Android App automation platform - PC controller"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "websockets>=12.0",
    "aiofiles>=23.2",
    "pyyaml>=6.0",
    "typer[all]>=0.9",
    "rich>=13.7",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
]
report = [
    "allure-pytest>=2.13",
    "jinja2>=3.1",
]
ocr = [
    "pillow>=10.2",
    "httpx>=0.27",
]
all = [
    "autotest[web,report,ocr]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
]

[project.scripts]
autotest = "autotest.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autotest"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
